服务器修改、优化测量

+++ embedded-projects/eis-system/user_code/lte.h
@@ -25,8 +25,8 @@ eis_status_t lte_module_init(void);
 eis_status_t lte_celluar_init(void);
 eis_status_t lte_connect_server(void);
 
-eis_status_t lte_http_get(const char *url, http_response_t *response, uint16_t timeout, uint8_t retry);
-eis_status_t lte_http_post(const char *url, const uint8_t *content, uint32_t content_size, http_response_t *response, uint16_t timeout, uint8_t retry);
+eis_status_t lte_http_get(const char *url, http_response_t *response, uint32_t timeout, uint8_t retry);
+eis_status_t lte_http_post(const char *url, const uint8_t *content, uint32_t content_size, http_response_t *response, uint32_t timeout, uint8_t retry);
 void response_purify(http_response_t *response);
 
 #endif /* USER_CODE_LTE_H_ */
